+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Details**
 - nflow-engine:
   - Only rollback poll operation when no workflows could be allocated for executing (when multiple pollers compete for same workflows)
+  - Allow configuring executor queue length with _nflow.dispatcher.executor.queue.size_
 
 ## 1.2.0 (2014-12-23)